requirements.txt
azure-communication-email ==1.0.*
pathutility == 1.0.*